Move resolver initialization to utils
To be re-used elsewhere in a future commit.
